% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cardinal.R
\name{cardinal}
\alias{cardinal}
\title{Cardinal Directions}
\usage{
cardinal(nr, nc, cdir = "north")
}
\arguments{
\item{nr}{number of rows (must be >=3)}

\item{nc}{number of columns (must be >=3)}

\item{cdir}{Direction character. Default "north".}
}
\value{
A matrix indicating, for each cell in a hypothetical 2D grid, the coordinates of the cell in the same matrix that lie in the specific cardinal direction.
}
\description{
Given a 2D grid of specified dimensions (i.e., matrix), returns indices of cells that lie in a cardinal orientation to each focal cell.
}
\details{
Goal of this function is to figure out which pairs of elements of a matrix (when listed linearly, e.g., a 3x3 matrix has elements 1-9, column-wise) lie a particular orientation to each other. These orientation are supplied via `cdir`, and are listed as the 4 cardinal directions (north west east south) and the 4 primary intercardinal directions (northwest southwest northeast southeast).
}

