% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colonize.R
\name{colonize}
\alias{colonize}
\title{Colonize}
\usage{
colonize(temps, spp.dens.temps, bio.mu, relative = FALSE, probs = NULL)
}
\arguments{
\item{temps}{}

\item{spp.dens.temps}{}

\item{bio.mu}{}

\item{relative}{Default FALSE; if TRUE, scale output so that maximum is 1}

\item{probs}{}
}
\value{
\code{suit.pers} is the probability that each element of \code{temps} is suitable for the species, i.e., the suitability of persistence (not a great name, I know)
\code{spp.pres} is NA or 1, where NA means the species is absent, and 1 means the species is present (rbinom(size=1, prob=suit.pers) realization); NA is used to be consistent with the conventions of the final element of the list, which is biomass, and which might be modeled on a \code{log()} scale
\code{spp.bio} are biomasses of the species. These biomasses are NA if the species is not present (i.e., an NA in spp.pres), and otherwise are equal to the \code{bio.mu} value associated with the species, adjusted up or down (additively) according to the probability of suitability. The adjustment is between -0.5 and 0.5, depending on p.suit. Biomasses are return as NA as opposed to 0 so that the biomass is unambiguous in the event that the bio.mu supplied is on a log() scale (exp(0)==1).
}
\description{
Colonize a grid with species given the tempereatures on the grid, the densities (from density() ) of those species across a wide range of temperatures, and the mean biomass of those species.
}
\details{
Function can be applied to 1 species or many species. If many, the length of bio.mu should be of length n.spp, where n.spp is the number of species, and spp.dens.temps should be a list of length n.spp of densities. \code{temps} would remain unchanged.
}

