% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dsample.R
\name{dsample}
\alias{dsample}
\title{Density Sample}
\usage{
dsample(ref, x, relative = FALSE)
}
\arguments{
\item{ref}{either an object of class density, or a numeric vector representing a random sample that should form the empirical reference distribution}

\item{x}{samples (quantiles) for which to calculate the density}

\item{relative}{Default FALSE, else calculation is scalled so that maximum returned is 1}
}
\description{
What is the density of each observation in X given the empirical distribution of a sample ref?
}
\details{
Similar to dnorm(), e.g., but instead of assuming the normal distribution, instead assumes an empirical distribution based on density(ref)
If ref is of class == "density", then the density of a value in x is calculated based on ref, otherwise, calculated based on density(ref).
}

