% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen.grid.R
\name{gen.grid}
\alias{gen.grid}
\title{Generate a Grid}
\usage{
gen.grid(x = NULL, dims = dim(x), byrow = FALSE, ...)
}
\arguments{
\item{x}{an array, matrix or numeric vector; NULL will fill with NA's}

\item{dims}{vector of length 3 specifying the integer size of dimensions}

\item{...}{arguments to be passed to \code{brick}}
}
\description{
Generate a grid that is a raster
}
\details{
Use it like array(), except it's a brick
}

