% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.cov.params.R
\name{get.cov.params}
\alias{get.cov.params}
\title{Get Covariate Parameters}
\usage{
get.cov.params(d)
}
\arguments{
\item{d}{the S.dens.X object from the guts of the simulation}
}
\description{
Helper function to get covariate parameters
}

