% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obs.chance.R
\name{obs.chance}
\alias{obs.chance}
\title{Observation Chance}
\usage{
obs.chance(dim2 = ns, dim1 = grid.t, dim3 = n.obs.reps,
  rand.gen = rnorm, chances, ...)
}
\arguments{
\item{dim2}{number of species; chances/ rng's always differ across \code{dim2} (unless chances supplied with 1 unique)}

\item{dim1}{time steps; chances will only differ across \code{dim1} if length of arguments in \code{...} are > 1}

\item{dim3}{replicates; chances will be same across \code{dim3}, but \code{dim1} shift by 1 index between each \code{dim3}}

\item{rand.gen}{a function for random number generation; its first argument must be \code{n}. Used to generate chances when \code{chances} is not specified.}

\item{chances}{Predetermined probabilities that can optionally be provided for each species. Also, length(chances) must be a multiple of \code{dim2.} If chances is supplied, \code{rand.gen} will not be used.}

\item{...}{arguments to be passed to \code{rand.gen}}
}
\value{
An array with observation chances, with first dimension varying time, second by species, and third by replicate.
}
\description{
# Helper function used to create time-varying chances of observation (detectability) for species
}
\details{
The main use of the function will be when chances is not supplied, and the function performs the random number generation. In a given year, all species will be drawn from the same distribution. That distribution can change among years by supplying vectors to \code{...}. Between replicates, the year that corresponds to a particular vectorized argument in \code{...} rotates. For example, if \code{rand.gen=rnorm}, and \code{mean=c(0:2)} is supplied and \code{dim1=3}, in the first replicate the first time step will have a mean of 0, the second of 1, and the third of 2. The second replicate the first year will have a mean of 2, the second of 0, and the third of 1.
}

