% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/obs.spp.R
\name{obs.spp}
\alias{obs.spp}
\title{Function for observing species}
\usage{
obs.spp(x, n.ss = 9, n.ss.mu, base.chance = rep(1, ns), t.noID)
}
\arguments{
\item{n.ss}{number of substrata that exist in each stratum (default 9). Must be a perfect square.}

\item{n.ss.mu}{number of substrata to sample over the whole region (default \code{trunc((n.ss*grid.w*grid.h)*(100/100))})}

\item{base.chance}{numeric vector with length equal number species indicating each species' detectability (0-1). Can leave it as missing.}
}
\description{
Given simulated species process, observe that process
}

