% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/p.perist.R
\name{p.persist}
\alias{p.persist}
\title{Probability of peristing}
\usage{
p.persist(temp.prop, temp.obs, method = c("ks.test", "dsample"))
}
\arguments{
\item{temp.prop}{a proposed temperature (environmental variable) in which to presist}

\item{temp.obs}{historically observed temperatures}

\item{method}{Either ks.test or dsample}
}
\description{
Given probaiblities over an environemntal gradient, and target environmental values, returns probability of persisting in the target
}

