% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_env.R
\name{plot_env}
\alias{plot_env}
\title{Plot Environment}
\usage{
plot_env(X, ...)
}
\arguments{
\item{X}{a spp object}

\item{...}{not used}
}
\value{
nothing at the moment
}
\description{
Plot simulated environment from a spp object
}
\details{
creates a plot of the environmental variable
}

