% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_env_bio.R
\name{plot_env_bio}
\alias{plot_env_bio}
\title{Plot Space-Time Environmental Variable and Biomass}
\usage{
plot_env_bio(X, n.steps = 6, spp.id, ...)
}
\arguments{
\item{X}{a spp object}

\item{n.steps}{number of time steps}

\item{spp.id}{integer (length 1) indicating index of species to plot; if missing picks one at random}

\item{...}{not used}
}
\description{
For one species at a time, and for a subset of the time steps, plot the biomass of that species in space
}

