% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_regRich_ts.R
\name{plot_regRich_ts}
\alias{plot_regRich_ts}
\title{Plot time series of true regional richness}
\usage{
plot_regRich_ts(X, ...)
}
\arguments{
\item{X}{a spp object}

\item{...}{not used}
}
\description{
Plot a simple time series of regional richness, meaning the richness plotted represented the aggregated number of unique species among all sites in each time step
}

