% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_rich_accum.R
\name{plot_rich_accum}
\alias{plot_rich_accum}
\title{Plot Richness Accumulation Curve for Each Year}
\usage{
plot_rich_accum(X, ...)
}
\arguments{
\item{X}{a spp object}

\item{...}{not used}
}
\description{
As you observe more places, you'll see more species. Plot that curve of species accumulated per sites visited for each year.
}

