% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_rich_st.R
\name{plot_rich_st}
\alias{plot_rich_st}
\title{Plot Richness over Space and Time}
\usage{
plot_rich_st(X, time = 1, ...)
}
\arguments{
\item{X}{a spp object}

\item{time}{time steps to plot; can be vector to plot multiple maps in same figure}
}
\description{
Plot the 2D map of richness, and make that map for each time step
}

