% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_simResp.R
\name{plot_simResp}
\alias{plot_simResp}
\title{Plot Simulated Response Curve}
\usage{
plot_simResp(X, ...)
}
\arguments{
\item{X}{a spp object}

\item{...}{not used}
}
\description{
Plot the probability of presence against the environmental variable for all species
}

