% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r2m.cell.R
\name{r2m.cell}
\alias{r2m.cell}
\title{Raster to matrix, by cell}
\usage{
r2m.cell(x, nrow = NULL, ncol = NULL, from = c("raster", "matrix"))
}
\arguments{
\item{x}{a raster or a matrix}

\item{nrow}{the rows}

\item{ncol}{the columns}

\item{from}{raster or matrix; whichever corresponds to x}
}
\description{
Provide the matrix to convert a raster order of elements to a matrix order
}
\details{
This is simply a quick index tool. Do not need to provide nrow and ncol, so long as they can be inferred by apply the corresponding function to \code{x}.
}

