% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.env.R
\name{sim.env}
\alias{sim.env}
\title{Simulate Environment}
\usage{
sim.env(grid.w = 5, grid.h = 7, grid.t = 12, X.slope = 0, h.slope = 1,
  half.val = 0, w.sd = 0.25, t.sd = 0.1)
}
\arguments{
\item{grid.w}{width of environmental grid (e.g., columns, longitude)}

\item{grid.h}{height of environmental grid (e.g., rows, latitude)}

\item{grid.t}{number of time steps for grid}

\item{X.slope}{the rate of change for the environmental variable across \code{grid.t}}

\item{h.slope}{increase in the environmental gradient per grid cell, going up}

\item{w.sd}{stanard deviation of environmental variable across the columns (\code{grid.w})}

\item{t.sd}{standard deviation of environmental variable between time steps}

\item{half.temp}{value of the environmental variable at the vertically middle grid cell in the middle time step}
}
\description{
Simulates the environmental gradient
}
\details{
Also introduces a small amount of noise to the temporal regression. Assumes that the environmental gradient varies across rows (\code{grid.h}), and adds noise within a row across columns. Increasing \code{h.slope} while holding \code{grid.h} constant will increase the range of values observed in the environmental variable; same is true for \code{X.slope} and \code{grid.t}.
}

