% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.spp.proc.R
\name{sim.spp.proc}
\alias{sim.spp.proc}
\title{Simulate Species Process}
\usage{
sim.spp.proc(grid.X, ns = 200, alpha_mu = c(0.5, 0, -0.025),
  alpha_sd = c(0.5, 0.21, 0.008), dynamic = FALSE, AR1.coef = 0.8,
  D.frac = 0.5, M.frac = 1, persist.bonus = 1)
}
\arguments{
\item{grid.X}{environmental grid}

\item{ns}{number of species}

\item{alpha_mu}{numeric vector of length 3, given the means of the intercept X coefficient, and X^2 coeff. X is the environmental varaible in grid.X.}

\item{alpha_sd}{the standard deviations associated with each of the terms described for \code{alpha_mu}}

\item{dynamic}{default TRUE means that species can only move to adjacent cells between years. FALSE means species aren't spatially restricted, and each year is indepenedent from the previous in terms of how many locations a species is occupying (independent given environmental grid)}

\item{AR1.coef}{If dynamic is TRUE, this is the autocorrelation coefficient of site and species-specific biomass between time steps}

\item{D.frac}{Dispersal fraction; if dynamic is TRUE, fraction of biomass to disperse from local cell to surroundings; think of it like reproduction, such that it doesn't result in a decrease in the biomass of the focal cell}

\item{M.frac}{Movement fraction; if dynamic is TRUE, proportion of biomass that will leave cell when it becomes unsuitable (this proportion is split evenly among neighbors, but not all neighbors will be suitable, so in many cases less than 100 percent of the biomass will find a new home)}

\item{perist.bonus}{Peristence bonus; if dynamic is TRUE, factor by which to adjust the probability that a species will fail to persist; decreasing this factors makes the species more likely to stick around, increasing it makes it less likely to stick around. When 1, no adjustment is made; when 0, the species will always stick around.}
}
\value{
An object of class spp and array
}
\description{
Simulate species occupying an environmental grid
}
\details{
Assumes the environmental grid is centered around 0. If X a matrix with an intercept the environmental variable, and the environmental variable squared, and Psi is probability of occupying X by species i, then \code{Psi = X\%*\%alpha}, where alpha has 3 parameters: a0, a1, a2, corresponding to an intercept and environmental coefficients. These values are 0.5, 0, and -0.025, respectively.
}

