% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_occ.R
\name{sim_occ}
\alias{sim_occ}
\title{Simulate Occupancy}
\usage{
sim_occ(..., detect.mus = 0, detect.sd = 0.1, format.msom = c("none",
  "jags", "stan"), n0s = 5, n.obs.reps = 1)
}
\arguments{
\item{...}{arguments to be passed to functions \code{\link{sim.env}}, \code{\link{sim.spp.proc}}, \code{\link{obs.spp}}.}

\item{detect.mus,detect.sd}{numeric vectors specifying the mean(s) and sd(s) of the parent distribution of species detectability on a logit scale. Default of 0 corresponds to 50\% detectability. Form detectabilities that are passed to t.noID \code{\link{obs.spp}}. Also, both can be of variable length, and need not be of same length, but must be multiples of each other, and it is best if lengths are multiples of \code{grid.t}.}

\item{format.msom}{character, format for msom? default is to simply output simulation}

\item{n0s}{if formatting for an msom, how many never-observed (0) species to pad with}

\item{n.obs.reps}{how many times to repeat the observation process, while holding the true process outcome the same, and holding the observation parameters the same (but perhaps changing which years get which detectability parameters)}
}
\value{
a list of spp objects
}
\description{
Simulate species occupying a 2D environmental grid, and the observation of that occupancy.
}
\details{
stuff
}
\examples{
sim_occ(ns=20)
}

