% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spp2msom.R
\name{spp2msom}
\alias{spp2msom}
\title{Convert spp to msom format}
\usage{
spp2msom(sim.obs)
}
\arguments{
\item{sim.obs}{observed simulation of spp, class spp}
}
\description{
Helper function to convert an spp object to a series of arrays that might be supplied to a model
}

