% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_sim_Bio.R
\name{plot_sim_bio_t}
\alias{plot_sim_bio_t}
\title{Plot Simulated Biomass at Time t}
\usage{
plot_sim_bio_t(X, time = 1, ...)
}
\arguments{
\item{X}{a spp object}

\item{time}{the time step}

\item{...}{not used}
}
\description{
Plot the spatial biomass of each species during the specified time step
}

